+++ noodles-vcf/CHANGELOG.md
@@ -10,6 +10,8 @@
 
   * vcf/header/format/key: Implement `Borrow<str>`.
 
+  * vcf/header/format/key: Add VCF 4.4 reserved keys.
+
   * vcf/header/info/key: Implement `Borrow<str>`.
 
   * vcf/header/info/key: Add VCF 4.4 reserved keys.